[BE #19]Refactor: 패키지 구조 개선 & 순환참조문제 해결 - 도메인 중심 패키지 - 인터페이스는 코어(auth, email..)에, 구현체는 auth/http, auth/store에 분리 - 모든 모듈이 구현체가 아닌 인터페이스만 참조하도록 변경
