+++ backend/internal/common/constants.go
@@ -1,0 +1,9 @@
+package common
+
+const (
+	OTPLength            = 6
+	MaxResendAttempts    = 5
+	ResendWaitSeconds    = 60
+	OTPExpirationMinutes = 10
+	TokenExpirationHours = 24
+)
