--- backend/internal/constants/auth.go
@@ -1,9 +1,0 @@
-package constants
-
-const (
-	OTPLength            = 6
-	MaxResendAttempts    = 5
-	ResendWaitSeconds    = 60
-	OTPExpirationMinutes = 10
-	TokenExpirationHours = 24
-)
