--- backend/internal/domain/model/database_status.go
@@ -1,16 +1,0 @@
-package model
-
-type InstanceStatus string
-
-const (
-	StatusProvisioning InstanceStatus = "provisioning"
-	StatusRunning      InstanceStatus = "running"
-	StatusStopped      InstanceStatus = "stopped"
-	StatusPaused       InstanceStatus = "paused" // 레몬 부족시
-	StatusError        InstanceStatus = "error"
-	StatusDeleting     InstanceStatus = "deleting"
-	StatusMaintenance  InstanceStatus = "maintenance"
-	StatusBackingUp    InstanceStatus = "backing_up"
-	StatusRestoring    InstanceStatus = "restoring"
-	StatusUpgrading    InstanceStatus = "upgrading"
-)
