--- backend/internal/domain/ports/secondary/session_repo.go
@@ -1,15 +1,0 @@
-package secondary
-
-import (
-	"context"
-
-	"github.com/piper-hyowon/dBtree/internal/domain/model"
-)
-
-type SessionRepo interface {
-	Save(ctx context.Context, session *model.AuthSession) error
-	GetByEmail(ctx context.Context, email string) (*model.AuthSession, error)
-	GetByToken(ctx context.Context, token string) (*model.AuthSession, error)
-	Delete(ctx context.Context, email string) error
-	Cleanup(ctx context.Context) error // 만료 세션 정리
-}
